Rename mod repo into repofile
