+++ src/main.rs
@@ -37,7 +37,7 @@ mod commands;
 mod crypto;
 mod id;
 mod index;
-mod repo;
+mod repofile;
 
 fn main() -> Result<()> {
     // this is a workaround until unix_sigpipe (https://github.com/rust-lang/rust/issues/97889) is available.
